# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PhotoSet 摄影套图浏览平台后端 API 文档\n提供套图管理、社区互动、用户认证、支付等功能",
        "title": "PhotoSet API",
        "termsOfService": "https://photoset.example.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@photoset.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/photosets": {
            "get": {
                "description": "获取套图列表，支持分页、标签、关键词等筛选条件（可选鉴权）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhotoSet"
                ],
                "summary": "获取套图列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认20，最大100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签名称",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "仅查看自己的作品",
                        "name": "mine",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态筛选(draft/published/pending)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新套图，需要登录（creator/admin角色）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhotoSet"
                ],
                "summary": "创建套图",
                "parameters": [
                    {
                        "description": "创建套图请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.PhotoSet"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/photosets/advanced": {
            "get": {
                "description": "高级搜索套图列表，支持分类、价格、排序等多维度筛选（可选鉴权）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhotoSet"
                ],
                "summary": "高级搜索套图",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认20，最大100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签名称",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "仅查看自己的作品",
                        "name": "mine",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态筛选(draft/published/pending)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类slug",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最低价格",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最高价格",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否免费",
                        "name": "is_free",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "时间范围",
                        "name": "time_range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户ID筛选",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/photosets/trash": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户已删除的套图列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhotoSet"
                ],
                "summary": "获取回收站列表",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.PhotoSet"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/photosets/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取套图的图片URL列表，需要登录认证",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhotoSet"
                ],
                "summary": "下载套图",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的套图ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权下载",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/photosets/{id}/permanent": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从数据库永久删除套图，不可恢复",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PhotoSet"
                ],
                "summary": "永久删除套图",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "永久删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的套图ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateRequest": {
            "type": "object",
            "required": [
                "cover",
                "title"
            ],
            "properties": {
                "category": {
                    "description": "分类 slug",
                    "type": "string"
                },
                "cover": {
                    "type": "string",
                    "maxLength": 500
                },
                "description": {
                    "type": "string"
                },
                "is_free": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Photo"
                    }
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "pending"
                    ]
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 200
                }
            }
        },
...
        "handlers.Photo": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "sort_order": {
                    "type": "integer"
                },
                "url": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "在请求头中添加 Bearer {token} 进行认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}